Fix race condition when cancelling a request
When cancelling a request pyFreenet immediately removed that request
from the job queue. It seem it was possible for the node to return the
requested data between just as pyFreenet cancels the request, thus
resulting in data being received that had no job.

This makes sure that the cancelled job is not removed from the queue
until the cancellation notification is received from the node.