Add option to specify the mime type for files matching a pattern
The option is saved per added site.

First time that this version is used, the option is saved for all sites.
This is not different from how other options are handled but could be
confusing because on subsequent updates, the option is ignored and the
value is set for all sites in your config-dir, even the ones you do not
update.