Fix usage summary for positional arguments
The previous way of argparse resulted in an incomplete and incorrect
usage summary, and was also misleading in source code because it looked
like the program was taking 1 or 2 file arguments, when the first
argument was always supposed to be a key.

The source seemed to be in two minds as to which of the two positional
arguments was actually optional. Most of it, including the comments,
seemed to assume it was the filename that was optional, but then the
code would take the one supplied positional argument and use it as a
filename.

I have now changed it so that it is the key that is optional and the
filename must always be supplied, giving "-" to read from stdin.

Reasoning: By my understanding, fcpupload is supposed to be a
simplified, easy to understand and easy to use equivalent to fcpput,
with a "just upload this" approach in mind. In this context, it seems to
be best to optimise the interface for the most likely usage: Upload a
given file to a deterministic CHK and return the key. Both specifying a
custom key and piping the data in via stdin are comparatively niche
applications.

Signed-off-by: Arne Babenhauserheide <arne_bab@web.de>